Only use the NoCloud data source
This is the only one which is required. Also load the isofs kernel
module automatically, so that cloud-init is able to mount the cloud-init
disk when NoCloud is the only source. For some reason, it was able to
do so with the full list of data sources. Some of them is probably using
"mout -t iso9660" and hence auto-loading isofs. And something else is
making NoCloud to pick this up after the fact. Weird.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>